+++ CHANGELOG.md
@@ -33,6 +33,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   planned to be removed in a future release.
   [#6016](https://github.com/jj-vcs/jj/issues/6016)
 
+* The `--allow-new` flag on `jj git push` is deprecated in favor of `--remote origin`.
+
+* The `git.push-new-bookmarks` config option is deprecated without replacement.
+  If there is only one remote, new bookmark can now be pushed by default without
+  this configuration option. If there is more than one remote, you must specify
+  `--remote` to push new bookmarks.
+
 ### New features
 
 * `jj workspace list` now accepts `-T`/`--template` option to customize its output via templates.
@@ -53,6 +60,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Add a `.parents()` method to the
   [`Operation`](docs/templates.md#operation-type) type in the templating
   language.
+
+* Pushing new bookmarks is now allowed by default if there is only one remote,
+  i.e. the flag `--allow-new` is not required anymore
 
 ### Fixed bugs